Fix uninitialized constant SimpleDelegator
Apparently, Ruby 2.7 requires that we require this library explicitly.
Although, it works in irb in Ruby 2.7.
Ruby 2.6 works without it.
Based on a PR I found for another project, it seems this problem also existed
in Ruby 2.2. So who knows.

That PR is here: https://github.com/jnunemaker/httparty/pull/532