Simplify user_create
This is a breaking change! This function was defined with a few special
qualities that I was not willing to add to the refactored system. These
include:
 - Validating which values are acceptable for "action".
 - Optionally requiring "password" if action is set to autoCreate.
 - Passing user info values in the root of the hash instead of inside a
   user_info hash.
This third item actually creates a breaking API change for users of
zoom_rb who use the user_create method. The actual API requires
this user_info hash, but this method deviated from the original API. I
think it is better for the Ruby API to track with the actual API. So
users of this library should update their code to wrap the user info
values in a user_info hash. This should be noted in the Changelog, as
this refactor is intended to be released as a major version release,
specifically version 1.0.