Fix OAuth bug
I made an error in my refactor and failed to change the headers for
OAuth requests. This will replace the "base_uri" option with a flag,
"oauth" which will automatically set the base_uri and the headers for
OAuth requests. The only requests that use these are Zoom::Actions::Token.